refactor: 예외 패키지 위치 변경, 다른 예외를 던지고 있던 문제 수정
